refactor: extract app runtime and session helpers
